pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gangtise-mcp-server"
version = "0.1.0"
description = "Gangtise 投研知识库 MCP Server"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
gangtise-mcp = "server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
